test(twenty-slack): split test helpers to one export per file
