+++ packages/twenty-apps/public/slack/src/__tests__/constants/slack-test-connected-account-id.constant.ts
@@ -1,2 +1,1 @@
-export const SLACK_TEST_WEBHOOK_SECRET = 'slack-test-signing-secret';
 export const SLACK_TEST_CONNECTED_ACCOUNT_ID = 'connected-account-slack-1';
