+++ packages/twenty-apps/public/slack/src/__tests__/constants/slack-test-webhook-secret.constant.ts
@@ -1,0 +1,1 @@
+export const SLACK_TEST_WEBHOOK_SECRET = 'slack-test-signing-secret';
