+++ packages/twenty-apps/public/slack/src/__tests__/slack-connection-lifecycle.integration-test.ts
@@ -1,14 +1,10 @@
 import { describe, expect, it } from 'vitest';
 
-import {
-  buildSlackAppHomeOpenedEventBody,
-  buildSlackMemberJoinedChannelEventBody,
-} from 'src/__tests__/utils/build-slack-event-payloads.util';
-import { SLACK_TEST_CONNECTED_ACCOUNT_ID } from 'src/__tests__/constants/slack-test.constants';
-import {
-  buildSlackAppConnection,
-  setupSlackIntegrationTest,
-} from 'src/__tests__/utils/setup-slack-integration-test.util';
+import { buildSlackAppHomeOpenedEventBody } from 'src/__tests__/utils/build-slack-app-home-opened-event-body.util';
+import { buildSlackMemberJoinedChannelEventBody } from 'src/__tests__/utils/build-slack-member-joined-channel-event-body.util';
+import { SLACK_TEST_CONNECTED_ACCOUNT_ID } from 'src/__tests__/constants/slack-test-connected-account-id.constant';
+import { buildSlackAppConnection } from 'src/__tests__/utils/build-slack-app-connection.util';
+import { setupSlackIntegrationTest } from 'src/__tests__/utils/setup-slack-integration-test.util';
 import { SLACK_BOT_USER_ID_KV_KEY } from 'src/logic-functions/constants/slack-bot-user-id-kv-key';
 import { SLACK_CHANNEL_WELCOME_TEXT } from 'src/logic-functions/constants/slack-channel-welcome-text';
 import { SLACK_CHANNEL_WELCOME_THREAD_TEXT } from 'src/logic-functions/constants/slack-channel-welcome-thread-text';
