+++ packages/twenty-apps/public/slack/src/__tests__/types/app-key-value-scope.type.ts
@@ -1,0 +1,1 @@
+export type AppKeyValueScope = 'WORKSPACE' | 'SERVER';
