+++ packages/twenty-apps/public/slack/src/__tests__/types/app-runtime-mock-options.type.ts
@@ -1,0 +1,7 @@
+import { type AppConnection } from 'twenty-sdk/logic-function';
+
+export type AppRuntimeMockOptions = {
+  apiUrl: string;
+  workspaceId: string;
+  connections?: AppConnection[];
+};
