+++ packages/twenty-apps/public/slack/src/__tests__/types/fake-slack-assistant-status.type.ts
@@ -1,0 +1,5 @@
+export type FakeSlackAssistantStatus = {
+  channelId: string;
+  threadTimestamp: string;
+  status: string;
+};
