+++ packages/twenty-apps/public/slack/src/__tests__/types/fake-slack-assistant-title.type.ts
@@ -1,0 +1,5 @@
+export type FakeSlackAssistantTitle = {
+  channelId: string;
+  threadTimestamp: string;
+  title: string;
+};
