+++ packages/twenty-apps/public/slack/src/__tests__/types/fake-slack-channel.type.ts
@@ -1,0 +1,11 @@
+export type FakeSlackChannel = {
+  id: string;
+  name: string;
+  isPrivate: boolean;
+  isArchived: boolean;
+  isMember: boolean;
+  isDirectMessage: boolean;
+  numMembers: number;
+  topic: string;
+  purpose: string;
+};
