+++ packages/twenty-apps/public/slack/src/__tests__/types/fake-slack-ephemeral-message.type.ts
@@ -1,0 +1,7 @@
+export type FakeSlackEphemeralMessage = {
+  channelId: string;
+  recipientUserId: string;
+  threadTimestamp?: string;
+  text?: string;
+  markdownText?: string;
+};
