+++ packages/twenty-apps/public/slack/src/__tests__/types/fake-slack-message.type.ts
@@ -1,0 +1,10 @@
+export type FakeSlackMessage = {
+  channelId: string;
+  timestamp: string;
+  threadTimestamp?: string;
+  userId?: string;
+  botId?: string;
+  text?: string;
+  markdownText?: string;
+  blocks?: unknown[];
+};
