+++ packages/twenty-apps/public/slack/src/__tests__/types/fake-slack-reaction.type.ts
@@ -1,0 +1,5 @@
+export type FakeSlackReaction = {
+  channelId: string;
+  messageTimestamp: string;
+  emojiName: string;
+};
