+++ packages/twenty-apps/public/slack/src/__tests__/types/fake-slack-suggested-prompts.type.ts
@@ -1,0 +1,5 @@
+export type FakeSlackSuggestedPrompts = {
+  channelId: string;
+  title: string;
+  prompts: unknown[];
+};
