+++ packages/twenty-apps/public/slack/src/__tests__/types/fake-slack-user.type.ts
@@ -1,0 +1,12 @@
+export type FakeSlackUser = {
+  id: string;
+  displayName?: string;
+  realName?: string;
+  email?: string;
+  teamId?: string;
+  isBot?: boolean;
+  isDeleted?: boolean;
+  isRestricted?: boolean;
+  isUltraRestricted?: boolean;
+  isEmailConfirmed?: boolean;
+};
