+++ packages/twenty-apps/public/slack/src/__tests__/types/slack-api-call.type.ts
@@ -1,0 +1,4 @@
+export type SlackApiCall = {
+  method: string;
+  args: Record<string, unknown>;
+};
