+++ packages/twenty-apps/public/slack/src/__tests__/types/slack-message-event-options.type.ts
@@ -1,0 +1,13 @@
+export type SlackMessageEventOptions = {
+  channelId: string;
+  text: string;
+  userId?: string;
+  messageTimestamp?: string;
+  threadTimestamp?: string;
+  channelType?: string;
+  eventId?: string;
+  teamId?: string;
+  botUserId?: string;
+  botId?: string;
+  subtype?: string;
+};
