+++ packages/twenty-apps/public/slack/src/__tests__/types/stored-key-value-entry.type.ts
@@ -1,0 +1,4 @@
+export type StoredKeyValueEntry = {
+  value: unknown;
+  workspaceId: string | null;
+};
